requirements.txt
# Web Framework
streamlit>=1.28.0

# Data Processing
pandas>=2.0.0
openpyxl>=3.1.0
pyarrow>=14.0.0

# Database and Analytics
duckdb>=0.9.0

# Visualization
plotly>=5.17.0

# Utilities
python-dotenv>=1.0.0
